invoice-extract/requirements.txt
fastapi
uvicorn[standard]
transformers>=4.42
accelerate
pillow
pymupdf
json5
torch
